Renamed deficit_plots.py to deficit_party_plots.py
