+++ deficit_party_plots.py
@@ -1,671 +1,354 @@
 '''
-This plots U.S. deficits by year against measures of political party prevalence
-in the Whitehouse and in Congress. It uses the stored dataset
-deficit_party_data.csv, the sources of which are documented in the first two
-rows of the .csv file. The variable names are in the fourth row and the data
-observations begin in the fifth row.
-
-This module defines the following function(s):
+The deficit_plots.py module prints plots of deficit/GDP for United States
+across years, by number of Democrat-held Senate seats and House seats, and by
+three different measures of party control. If a user runs this module as a
+script, it will create all six plots.
 '''
 
 # Import packages
+from bokeh.core.property.numeric import Interval
+from bokeh.io.output import reset_output
+from bokeh.models.annotations import Label, LabelSet
+from bokeh.models.glyphs import VArea
+from bokeh.models.layouts import Panel
+from bokeh.models.tickers import SingleIntervalTicker
 import numpy as np
 import pandas as pd
 import datetime as dt
 import os
-from bokeh.io import output_file
+from bokeh.io import output_file, save
 from bokeh.plotting import figure, show
-from bokeh.models import ColumnDataSource, Title, Legend, HoverTool
-# from bokeh.models import Label
-# from bokeh.palettes import Category20
+from bokeh.models import (ColumnDataSource, CDSView, GroupFilter, Title,
+                          Legend, HoverTool, NumeralTickFormatter, Span, Tabs,
+                          Panel)
 
-'''
+# Set paths to work across Mac/Windows/Linux platforms
 cur_path = os.path.split(os.path.abspath(__file__))[0]
-filename_data = ('data/deficit_party_data.csv')
-data_file_path = os.path.join(cur_path, filename_data)
-alldata_df = pd.read_csv(
-    data_file_path,
-    names=['Year', 'deficit_gdp', 'president', 'president_party',
-           'congress_num', 'congress_sess', 'dem_whitehouse',
-           'dem_senateseats', 'rep_senateseats', 'oth_senateseats',
-           'tot_senateseats', 'dem_houseseats', 'rep_houseseats',
-           'oth_houseseats', 'tot_houseseats'],
-    dtype = {'Year': 'int64', 'deficit_gdp': 'float64',
-             'president': 'str', 'president_party': 'str',
-             'congress_num': 'int64', 'congress_sess': 'int64',
-             'dem_whitehouse': 'int64', 'dem_senateseats': 'int64',
-             'rep_senateseats': 'int64', 'oth_senateseats': 'int64',
-             'tot_senateseats': 'int64', 'dem_houseseats': 'int64',
-             'rep_houseseats': 'int64', 'oth_houseseats': 'int64',
-             'tot_houseseats': 'int64'},
-    parse_dates=['Year'], skiprows=4)
-'''
-
-#Reading data from CVS (deficit_party_data.csv)
-alldata_df = pd.read_csv(    'data\deficit_party_data.csv',
-                            dtype = {     'Year': np.int64, 
-                                          'deficit_gdp': np.float64, 
-                                          'president': 'str', 
-                                          'president_party': 'str',
-                                          'congress_num': np.int64, 
-                                          'congress_sess': np.int64,
-                                          'dem_whitehouse': np.int64, 
-                                          'dem_senateseats': np.int64,
-                                          'rep_senateseats': np.int64, 
-                                          'oth_senateseats': np.int64,
-                                          'tot_senateseats': np.int64, 
-                                          'dem_houseseats': np.int64,
-                                          'rep_houseseats': np.int64, 
-                                          'oth_houseseats': np.int64,
-                                          'tot_houseseats': np.int64},
-                            skiprows=3)
-#print(alldata_df)
-deficit_party_cds = ColumnDataSource(alldata_df)
-
-# Create Bokeh plot of Deficits/GDP by Democrat-held House seats
-fig_title = 'U.S. Deficits/GDP by Democrat-held House Seats: 1929-2020'
-filename = ('images/deficitimage_senate_demseats.html')
-output_file(filename, title=fig_title)
-
-# Format the tooltip
-tooltips = [('Year', '@Year{0.}'),
-            ('Deficit/GDP', '@deficit_gdp{0.0}'+'%'),
-            ('Democrat Seats', '@DemHouseSeats{0.}')]
-
-
-# Solve for minimum and maximum Democrat House seats and deficit/GDP for
-# plotting purposes
-min_demseats_val = alldata_df['DemHouseSeats'].min()
-print('Minimum Democrat House seats:', min_demseats_val)
-max_demseats_val = alldata_df['DemHouseSeats'].max()
-print('Maximum Democrat House seats:', max_demseats_val)
-min_defGDP_val = alldata_df['deficit_gdp'].min()
-print('Minimum deficit/GDP:', min_defGDP_val)
-max_defGDP_val = alldata_df['deficit_gdp'].max()
-print('Maximum deficit/GDP:', max_defGDP_val)
-
-datarange_demseats_vals = max_demseats_val - min_demseats_val
-datarange_defGDP_vals = max_defGDP_val - min_defGDP_val
-fig_buffer_pct = 0.10
-fig = figure(plot_height=500,
-             plot_width=800,
-             x_axis_label='Number of Democratic House Seats (out of 435)',
-             y_axis_label='Deficit/GDP',
-             y_range=(min_defGDP_val - fig_buffer_pct * datarange_defGDP_vals,
-                      max_defGDP_val + fig_buffer_pct * datarange_defGDP_vals),
-             x_range=((min_demseats_val -
-                       fig_buffer_pct * datarange_demseats_vals),
-                      (max_demseats_val +
-                       fig_buffer_pct * datarange_demseats_vals)),
-             tools=['save', 'zoom_in', 'zoom_out', 'box_zoom',
-                    'pan', 'undo', 'redo', 'reset', 'help'],
-             toolbar_location='left')
-
-deficit_house_df = alldata_df[['Year', 'deficit_gdp', 'DemHouseSeats']]
-deficit_house_cds = ColumnDataSource(deficit_house_df)
-fig.title.text_font_size = '18pt'
-fig.toolbar.logo = None
-fig.scatter(x='DemHouseSeats', y='deficit_gdp', source=deficit_house_df,
-            line_color='black', fill_color='blue', size=12, alpha=0.7)
-fig.add_tools(HoverTool(tooltips=tooltips))
-
-# Dashed vertical line at 217 House seats representing the 50% mark
-fig.line(x=[217, 217], y=[-30, 10], color='black', line_width=2,
-         line_dash='dashed', alpha=0.5)
-
-# Dashed horizontal line at deficit/GDP = 0
-fig.line(x=[0, 435], y=[0.0, 0.0], color='black', line_width=2,
-         line_dash='dashed', alpha=0.5)
-
-# Create the tick marks for the x-axis and set x-axis labels
-major_tick_labels = ['150', '200', '217', '250', '300', '350']
-major_tick_list = [150, 200, 217, 250, 300, 350]
-
-minor_tick_list = [item for item in range(150, 351, 10)]
-major_tick_dict = dict(zip(major_tick_list, major_tick_labels))
-fig.xaxis.ticker = major_tick_list
-fig.xaxis.major_label_overrides = major_tick_dict
-
-show(fig)
-
-# # Add legend
-# legend = Legend(items=[(rec_label_yrmth_lst[0], [l0]),
-#                         (rec_label_yrmth_lst[1], [l1]),
-#                         (rec_label_yrmth_lst[2], [l2]),
-#                         (rec_label_yrmth_lst[3], [l3]),
-#                         (rec_label_yrmth_lst[4], [l4]),
-#                         (rec_label_yrmth_lst[5], [l5]),
-#                         (rec_label_yrmth_lst[6], [l6]),
-#                         (rec_label_yrmth_lst[7], [l7]),
-#                         (rec_label_yrmth_lst[8], [l8]),
-#                         (rec_label_yrmth_lst[9], [l9]),
-#                         (rec_label_yrmth_lst[10], [l10]),
-#                         (rec_label_yrmth_lst[11], [l11]),
-#                         (rec_label_yrmth_lst[12], [l12]),
-#                         (rec_label_yrmth_lst[13], [l13]),
-#                         (rec_label_yrmth_lst[14], [l14])],
-#                 location='center')
-# fig.add_layout(legend, 'right')
-
-# # # Add label to current recession low point
-# # fig.text(x=[12, 12, 12, 12], y=[0.63, 0.60, 0.57, 0.54],
-# #          text=['2020-03-23', 'DJIA: 18,591.93', '63.3% of peak',
-# #                '39 days from peak'],
-# #          text_font_size='8pt', angle=0)
-
-# # label_text = ('Recent low \n 2020-03-23 \n DJIA: 18,591.93 \n '
-# #               '63\% of peak \n 39 days from peak')
-# # fig.add_layout(Label(x=10, y=0.65, x_units='screen', text=label_text,
-# #                      render_mode='css', border_line_color='black',
-# #                      border_line_alpha=1.0,
-# #                      background_fill_color='white',
-# #                      background_fill_alpha=1.0))
-
-# # Add title and subtitle to the plot
-# fig_title2 = 'Progression of U.S. total nonfarm employment'
-# fig_title3 = '(PAYEMS, seasonally adjusted) in last 15 recessions'
-# fig.add_layout(Title(text=fig_title3, text_font_style='bold',
-#                         text_font_size='16pt', align='center'), 'above')
-# fig.add_layout(Title(text=fig_title2, text_font_style='bold',
-#                         text_font_size='16pt', align='center'), 'above')
-
-# # Add source text below figure
-# updated_date_str = end_date.strftime('%B %-d, %Y')
-# fig.add_layout(Title(text='Source: Richard W. Evans (@RickEcon), ' +
-#                             'historical PAYEMS data from FRED and BLS, ' +
-#                             'updated ' + updated_date_str + '.',
-#                         align='left',
-#                         text_font_size='3mm',
-#                         text_font_style='italic'),
-#                 'below')
-# fig.legend.click_policy = 'mute'
-
-# # Add the HoverTool to the figure
-# fig.add_tools(HoverTool(tooltips=tooltips, toggleable=False,
-#                         formatters={'@Date': 'datetime'}))
-
-# if html_show:
-#show(fig)
-
-
-
-# def get_usempl_data(frwd_mths_max, bkwd_mths_max, end_date_str,
-#                     download_from_internet=True):
-#     '''
-#     This function either downloads or reads in the U.S. total nonfarm payrolls
-#     seasonally adjusted monthly data series (PAYEMS) and adds variables
-#     mths_frm_peak and empl_dv_pk for each of the last 15 recessions.
-
-#     Args:
-#         frwd_mths_max (int): maximum number of months forward from the peak
-#             month to plot
-#         bckwd_mths_max (int): maximum number of months backward from the peak
-#             month to plot
-#         end_date_str (str): end date of PAYEMS time series in 'YYYY-mm-dd'
-#             format
-#         download_from_internet (bool): =True if download data from
-#             fred.stlouisfed.org, otherwise read data in from local directory
-
-#     Other functions and files called by this function:
-#         usempl_[yyyy-mm-dd].csv
-
-#     Files created by this function:
-#         usempl_[yyyy-mm-dd].csv
-#         usempl_pk_[yyyy-mm-dd].csv
-
-#     Returns:
-#         usempl_pk (DataFrame): N x 46 DataFrame of mths_frm_peak, Date{i},
-#             Close{i}, and close_dv_pk{i} for each of the 15 recessions for the
-#             periods specified by bkwd_days_max and frwd_days_max
-#         end_date_str2 (str): actual end date of DJIA time series in
-#             'YYYY-mm-dd' format. Can differ from the end_date input to this
-#             function if the final data for that day have not come out yet
-#             (usually 2 hours after markets close, 6:30pm EST), or if the
-#             end_date is one on which markets are closed (e.g. weekends and
-#             holidays). In this latter case, the pandas_datareader library
-#             chooses the most recent date for which we have DJIA data.
-#         peak_vals (list): list of peak DJIA value at the beginning of each of
-#             the last 15 recessions
-#         peak_dates (list): list of string date (YYYY-mm-dd) of peak DJIA value
-#             at the beginning of each of the last 15 recessions
-#         rec_label_yr_lst (list): list of string start year and end year of each
-#             of the last 15 recessions
-#         rec_label_yrmth_lst (list): list of string start year and month and end
-#             year and month of each of the last 15 recessions
-#         rec_beg_yrmth_lst (list): list of string start year and month of each
-#             of the last 15 recessions
-#         maxdate_rng_lst (list): list of tuples with start string date and end
-#             string date within which range we define the peak DJIA value at the
-#             beginning of each of the last 15 recessions
-#     '''
-#     end_date = dt.datetime.strptime(end_date_str, '%Y-%m-%d')
-
-#     # Name the current directory and make sure it has a data folder
-#     cur_path = os.path.split(os.path.abspath(__file__))[0]
-#     data_fldr = 'data'
-#     data_dir = os.path.join(cur_path, data_fldr)
-#     if not os.access(data_dir, os.F_OK):
-#         os.makedirs(data_dir)
-
-#     filename_basic = ('data/usempl_' + end_date_str + '.csv')
-#     filename_full = ('data/usempl_pk_' + end_date_str + '.csv')
-
-#     if download_from_internet:
-#         # Download the employment data directly from fred.stlouisfed.org
-#         # (requires internet connection)
-#         start_date = dt.datetime(1939, 1, 1)
-#         usempl_df = pddr.fred.FredReader(symbols='PAYEMS', start=start_date,
-#                                          end=end_date).read()
-#         usempl_df = pd.DataFrame(usempl_df).sort_index()  # Sort old to new
-#         usempl_df = usempl_df.reset_index(level=['DATE'])
-#         usempl_df = usempl_df.rename(columns={'DATE': 'Date'})
-#         end_date_str2 = usempl_df['Date'].iloc[-1].strftime('%Y-%m-%d')
-#         end_date = dt.datetime.strptime(end_date_str2, '%Y-%m-%d')
-#         filename_basic = ('data/usempl_' + end_date_str2 + '.csv')
-#         filename_full = ('data/usempl_pk_' + end_date_str2 + '.csv')
-#         usempl_df.to_csv(filename_basic, index=False)
-#         # Merge in U.S. annual average nonfarm payroll employment (not
-#         # seasonally adjusted) 1919-1938. Date values for annual data are set
-#         # to July 1 of that year. These data are taken from Table 1 on page 1
-#         # of "Employment, Hours, and Earnings, United States, 1909-90, Volume
-#         # I," Bulletin of the United States Bureau of Labor Statistics, No.
-#         # 2370, March 1991.
-#         # <https://fraser.stlouisfed.org/title/employment-earnings-united-
-#         # states-189/employment-hours-earnings-united-states-1909-90-5435/
-#         # content/pdf/emp_bmark_1909_1990_v1>
-#         filename_annual = ('data/usempl_anual_1919-1938.csv')
-#         ann_data_file_path = os.path.join(cur_path, filename_annual)
-#         usempl_ann_df = pd.read_csv(ann_data_file_path,
-#                                     names=['Date', 'PAYEMS'],
-#                                     parse_dates=['Date'], skiprows=1,
-#                                     na_values=['.', 'na', 'NaN'])
-#         usempl_df = usempl_df.append(usempl_ann_df, ignore_index=True)
-#         usempl_df = usempl_df.sort_values(by='Date')
-#         usempl_df = usempl_df.reset_index(drop=True)
-#         usempl_df.to_csv(filename_basic, index=False)
-#         # Add other months to annual data 1919-01-01 to 1938-12-01 and fill in
-#         # artificial employment data by cubic spline interpolation
-#         months_df = \
-#             pd.DataFrame(pd.date_range('1919-01-01', '1938-12-01', freq='MS'),
-#                          columns=['Date'])
-#         usempl_df = pd.merge(usempl_df, months_df, left_on='Date',
-#                              right_on='Date', how='outer')
-#         usempl_df = usempl_df.sort_values(by='Date')
-#         usempl_df = usempl_df.reset_index(drop=True)
-#         usempl_df['PAYEMS'].iloc[:242] = \
-#             usempl_df['PAYEMS'].iloc[:242].interpolate(method='cubic')
-#     else:
-#         # Import the data as pandas DataFrame
-#         end_date_str2 = end_date_str
-#         data_file_path = os.path.join(cur_path, filename_basic)
-#         usempl_df = pd.read_csv(data_file_path, names=['Date', 'PAYEMS'],
-#                                 parse_dates=['Date'], skiprows=1,
-#                                 na_values=['.', 'na', 'NaN'])
-#         usempl_df = usempl_df.dropna()
-
-#     print('End date of U.S. employment series is',
-#           end_date.strftime('%Y-%m-%d'))
-
-#     # Set recession-specific parameters
-#     rec_label_yr_lst = \
-#         ['1929-1933',  # (Aug 1929 - Mar 1933) Great Depression
-#          '1937-1938',  # (May 1937 - Jun 1938)
-#          '1945',       # (Feb 1945 - Oct 1945)
-#          '1948-1949',  # (Nov 1948 - Oct 1949)
-#          '1953-1954',  # (Jul 1953 - May 1954)
-#          '1957-1958',  # (Aug 1957 - Apr 1958)
-#          '1960-1961',  # (Apr 1960 - Feb 1961)
-#          '1969-1970',  # (Dec 1969 - Nov 1970)
-#          '1973-1975',  # (Nov 1973 - Mar 1975)
-#          '1980',       # (Jan 1980 - Jul 1980)
-#          '1981-1982',  # (Jul 1981 - Nov 1982)
-#          '1990-1991',  # (Jul 1990 - Mar 1991)
-#          '2001',       # (Mar 2001 - Nov 2001)
-#          '2007-2009',  # (Dec 2007 - Jun 2009) Great Recession
-#          '2020-pres']  # (Feb 2020 - present) Coronavirus recession
-
-#     rec_label_yrmth_lst = ['Aug 1929 - Mar 1933',  # Great Depression
-#                            'May 1937 - Jun 1938',
-#                            'Feb 1945 - Oct 1945',
-#                            'Nov 1948 - Oct 1949',
-#                            'Jul 1953 - May 1954',
-#                            'Aug 1957 - Apr 1958',
-#                            'Apr 1960 - Feb 1961',
-#                            'Dec 1969 - Nov 1970',
-#                            'Nov 1973 - Mar 1975',
-#                            'Jan 1980 - Jul 1980',
-#                            'Jul 1981 - Nov 1982',
-#                            'Jul 1990 - Mar 1991',
-#                            'Mar 2001 - Nov 2001',
-#                            'Dec 2007 - Jun 2009',  # Great Recession
-#                            'Feb 2020 - present']  # Coronavirus recess'n
-
-#     rec_beg_yrmth_lst = ['Aug 1929', 'May 1937', 'Feb 1945', 'Nov 1948',
-#                          'Jul 1953', 'Aug 1957', 'Apr 1960', 'Dec 1969',
-#                          'Nov 1973', 'Jan 1980', 'Jul 1981', 'Jul 1990',
-#                          'Mar 2001', 'Dec 2007', 'Feb 2020']
-
-#     maxdate_rng_lst = [('1929-7-1', '1929-10-1'),
-#                        ('1937-7-1', '1937-7-1'),
-#                        ('1945-1-1', '1945-3-1'),
-#                        ('1948-9-1', '1949-1-1'),
-#                        ('1953-6-1', '1953-8-1'),
-#                        ('1957-7-1', '1957-9-1'),
-#                        ('1960-3-1', '1960-5-1'),
-#                        ('1969-11-1', '1970-3-1'),
-#                        ('1973-10-1', '1974-7-1'),
-#                        ('1979-12-1', '1980-3-1'),
-#                        ('1981-6-1', '1981-8-1'),
-#                        ('1990-6-1', '1991-8-1'),
-#                        ('2001-2-1', '2001-4-1'),
-#                        ('2007-11-1', '2008-1-1'),
-#                        ('2020-1-1', '2020-3-1')]
-
-#     # Create normalized peak series for each recession
-#     usempl_pk = \
-#         pd.DataFrame(np.arange(-bkwd_mths_max, frwd_mths_max + 1, dtype=int),
-#                      columns=['mths_frm_peak'])
-#     usempl_pk_long = usempl_df.copy()
-#     peak_vals = []
-#     peak_dates = []
-#     for i, maxdate_rng in enumerate(maxdate_rng_lst):
-#         # Identify peak closing value within two months (with only 2
-#         # exceptions) of the beginning month of the recession
-#         peak_val = \
-#             usempl_df['PAYEMS'][(usempl_df['Date'] >= maxdate_rng[0]) &
-#                                 (usempl_df['Date'] <= maxdate_rng[1])].max()
-#         peak_vals.append(peak_val)
-#         usempl_dv_pk_name = 'usempl_dv_pk' + str(i)
-#         usempl_pk_long[usempl_dv_pk_name] = (usempl_pk_long['PAYEMS'] /
-#                                              peak_val)
-#         # Identify date of peak PAYEMS value within two months (with
-#         # only 2 exceptions) of the beginning month of the recession
-#         peak_date = \
-#             usempl_df['Date'][(usempl_df['Date'] >= maxdate_rng[0]) &
-#                               (usempl_df['Date'] <= maxdate_rng[1]) &
-#                               (usempl_df['PAYEMS'] == peak_val)].max()
-#         peak_dates.append(peak_date.strftime('%Y-%m-%d'))
-#         mths_frm_pk_name = 'mths_frm_pk' + str(i)
-#         usempl_pk_long[mths_frm_pk_name] = \
-#             ((usempl_pk_long['Date'].dt.year - peak_date.year) * 12 +
-#              (usempl_pk_long['Date'].dt.month - peak_date.month))
-#         # usempl_pk_long[mths_frm_pk_name] = (usempl_pk_long['Date'] -
-#         #                                     peak_date).dt.years
-#         print('peak_val ' + str(i) + ' is', peak_val, 'on date',
-#               peak_date.strftime('%Y-%m-%d'), '(Beg. rec. month:',
-#               rec_beg_yrmth_lst[i], ')')
-#         # I need to merge the data into this new usempl_pk DataFrame so that
-#         # mths_frm_peak variable is shared across the dataframe
-#         usempl_pk = \
-#             pd.merge(usempl_pk,
-#                      usempl_pk_long[[mths_frm_pk_name, 'Date', 'PAYEMS',
-#                                      usempl_dv_pk_name]],
-#                      left_on='mths_frm_peak', right_on=mths_frm_pk_name,
-#                      how='left')
-#         usempl_pk.drop(columns=[mths_frm_pk_name], inplace=True)
-#         usempl_pk.rename(
-#             columns={'Date': f'Date{i}', 'PAYEMS': f'PAYEMS{i}'}, inplace=True)
-
-#     usempl_pk.to_csv(filename_full, index=False)
-
-#     return (usempl_pk, end_date_str2, peak_vals, peak_dates, rec_label_yr_lst,
-#             rec_label_yrmth_lst, rec_beg_yrmth_lst, maxdate_rng_lst)
-
-
-# def usempl_npp(frwd_mths_main=12, bkwd_mths_main=2, frwd_mths_max=135,
-#                bkwd_mths_max=48, usempl_end_date='today',
-#                download_from_internet=True, html_show=True):
-#     '''
-#     This function creates the HTML and JavaScript code for the dynamic
-#     visualization of the normalized peak plot of the last 15 recessions in the
-#     United States, from the Great Depression (Aug. 1929 - Mar. 1933) to the
-#     most recent COVID-19 recession (Feb. 2020 - present).
-
-#     Args:
-#         frwd_mths_main (int): number of months forward from the peak to plot in
-#             the default main window of the visualization
-#         bkwd_mths_maim (int): number of months backward from the peak to plot
-#             in the default main window of the visualization
-#         frwd_mths_max (int): maximum number of months forward from the peak to
-#             allow for the plot, to be seen by zooming out
-#         bkwd_mths_max (int): maximum number of months backward from the peak to
-#             allow for the plot, to be seen by zooming out
-#         usempl_end_date (str): either 'today' or the end date of PAYEMS time
-#             series in 'YYYY-mm-dd' format
-#         download_from_internet (bool): =True if download data from St. Louis
-#             Federal Reserve's FRED system
-#             (https://fred.stlouisfed.org/series/PAYEMS), otherwise read data in
-#             from local directory
-#         html_show (bool): =True if open dynamic visualization in browser once
-#             created
-
-#     Other functions and files called by this function:
-#         get_usempl_data()
-
-#     Files created by this function:
-#        images/usempl_[yyyy-mm-dd].html
-
-#     Returns: fig, end_date_str
-#     '''
-#     # Create directory if images directory does not already exist
-#     cur_path = os.path.split(os.path.abspath(__file__))[0]
-#     image_fldr = 'images'
-#     image_dir = os.path.join(cur_path, image_fldr)
-#     if not os.access(image_dir, os.F_OK):
-#         os.makedirs(image_dir)
-
-#     if usempl_end_date == 'today':
-#         end_date = dt.date.today()  # Go through today
-#     else:
-#         end_date = dt.datetime.strptime(usempl_end_date, '%Y-%m-%d')
-
-#     end_date_str = end_date.strftime('%Y-%m-%d')
-
-#     # Set main window and total data limits for monthly plot
-#     frwd_mths_main = int(frwd_mths_main)
-#     bkwd_mths_main = int(bkwd_mths_main)
-#     frwd_mths_max = int(frwd_mths_max)
-#     bkwd_mths_max = int(bkwd_mths_max)
-
-#     (usempl_pk, end_date_str2, peak_vals, peak_dates, rec_label_yr_lst,
-#         rec_label_yrmth_lst, rec_beg_yrmth_lst, maxdate_rng_lst) = \
-#         get_usempl_data(frwd_mths_max, bkwd_mths_max, end_date_str,
-#                         download_from_internet)
-#     if end_date_str2 != end_date_str:
-#         print('PAYEMS data downloaded on ' + end_date_str + ' has most ' +
-#               'recent PAYEMS data month of ' + end_date_str2 + '.')
-#     end_date2 = dt.datetime.strptime(end_date_str2, '%Y-%m-%d')
-
-#     rec_cds_list = []
-#     min_main_val_lst = []
-#     max_main_val_lst = []
-#     for i in range(15):
-#         usempl_pk_rec = \
-#             usempl_pk[['mths_frm_peak', f'Date{i}', f'PAYEMS{i}',
-#                        f'usempl_dv_pk{i}']].dropna()
-#         usempl_pk_rec.rename(
-#             columns={f'Date{i}': 'Date', f'PAYEMS{i}': 'PAYEMS',
-#                      f'usempl_dv_pk{i}': 'usempl_dv_pk'}, inplace=True)
-#         rec_cds_list.append(ColumnDataSource(usempl_pk_rec))
-#         # Find minimum and maximum usempl_dv_pk values as inputs to main plot
-#         # frame size
-#         min_main_val_lst.append(
-#             usempl_pk_rec['usempl_dv_pk'][
-#                 (usempl_pk_rec['mths_frm_peak'] >= -bkwd_mths_main) &
-#                 (usempl_pk_rec['mths_frm_peak'] <= frwd_mths_main)].min())
-#         max_main_val_lst.append(
-#             usempl_pk_rec['usempl_dv_pk'][
-#                 (usempl_pk_rec['mths_frm_peak'] >= -bkwd_mths_main) &
-#                 (usempl_pk_rec['mths_frm_peak'] <= frwd_mths_main)].max())
-
-#     # Create Bokeh plot of PAYEMS normalized peak plot figure
-#     fig_title = 'Progression of PAYEMS in last 15 recessions'
-#     filename = ('images/usempl_npp_' + end_date_str2 + '.html')
-#     output_file(filename, title=fig_title)
-
-#     # Format the tooltip
-#     tooltips = [('Date', '@Date{%F}'),
-#                 ('Months from peak', '$x{0.}'),
-#                 ('Employment', '@PAYEMS{0,0.},000'),
-#                 ('Fraction of peak', '@usempl_dv_pk{0.0 %}')]
-
-#     # Solve for minimum and maximum PAYEMS/Peak values in monthly main display
-#     # window in order to set the appropriate xrange and yrange
-#     min_main_val = min(min_main_val_lst)
-#     max_main_val = max(max_main_val_lst)
-
-#     datarange_main_vals = max_main_val - min_main_val
-#     datarange_main_mths = int(frwd_mths_main + bkwd_mths_main)
-#     fig_buffer_pct = 0.10
-#     fig = figure(plot_height=500,
-#                  plot_width=800,
-#                  x_axis_label='Months from Peak',
-#                  y_axis_label='PAYEMS as fraction of Peak',
-#                  y_range=(min_main_val - fig_buffer_pct * datarange_main_vals,
-#                           max_main_val + fig_buffer_pct * datarange_main_vals),
-#                  x_range=((-bkwd_mths_main -
-#                            fig_buffer_pct * datarange_main_mths),
-#                           (frwd_mths_main +
-#                            fig_buffer_pct * datarange_main_mths)),
-#                  tools=['save', 'zoom_in', 'zoom_out', 'box_zoom',
-#                         'pan', 'undo', 'redo', 'reset', 'hover', 'help'],
-#                  toolbar_location='left')
-#     fig.title.text_font_size = '18pt'
-#     fig.toolbar.logo = None
-#     l0 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[0],
-#                   color='blue', line_width=5, alpha=0.7, muted_alpha=0.15)
-#     l1 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[1],
-#                   color=Category20[13][0], line_width=2, alpha=0.7,
-#                   muted_alpha=0.15)
-#     l2 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[2],
-#                   color=Category20[13][1], line_width=2, alpha=0.7,
-#                   muted_alpha=0.15)
-#     l3 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[3],
-#                   color=Category20[13][2], line_width=2,
-#                   alpha=0.7, muted_alpha=0.15)
-#     l4 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[4],
-#                   color=Category20[13][3], line_width=2, alpha=0.7,
-#                   muted_alpha=0.15)
-#     l5 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[5],
-#                   color=Category20[13][4], line_width=2, alpha=0.7,
-#                   muted_alpha=0.15)
-#     l6 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[6],
-#                   color=Category20[13][5], line_width=2, alpha=0.7,
-#                   muted_alpha=0.15)
-#     l7 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[7],
-#                   color=Category20[13][6], line_width=2, alpha=0.7,
-#                   muted_alpha=0.15)
-#     l8 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[8],
-#                   color=Category20[13][7], line_width=2, alpha=0.7,
-#                   muted_alpha=0.15)
-#     l9 = fig.line(x='mths_frm_peak', y='usempl_dv_pk', source=rec_cds_list[9],
-#                   color=Category20[13][8], line_width=2, alpha=0.7,
-#                   muted_alpha=0.15)
-#     l10 = fig.line(x='mths_frm_peak', y='usempl_dv_pk',
-#                    source=rec_cds_list[10], color=Category20[13][9],
-#                    line_width=2, alpha=0.7, muted_alpha=0.15)
-#     l11 = fig.line(x='mths_frm_peak', y='usempl_dv_pk',
-#                    source=rec_cds_list[11], color=Category20[13][10],
-#                    line_width=2, alpha=0.7, muted_alpha=0.15)
-#     l12 = fig.line(x='mths_frm_peak', y='usempl_dv_pk',
-#                    source=rec_cds_list[12], color=Category20[13][11],
-#                    line_width=2, alpha=0.7, muted_alpha=0.15)
-#     l13 = fig.line(x='mths_frm_peak', y='usempl_dv_pk',
-#                    source=rec_cds_list[13], color=Category20[13][12],
-#                    line_width=2, alpha=0.7, muted_alpha=0.15)
-#     l14 = fig.line(x='mths_frm_peak', y='usempl_dv_pk',
-#                    source=rec_cds_list[14], color='black', line_width=5,
-#                    alpha=0.7, muted_alpha=0.15)
-
-#     # Dashed vertical line at the peak PAYEMS value period
-#     fig.line(x=[0.0, 0.0], y=[-0.5, 2.0], color='black', line_width=2,
-#              line_dash='dashed', alpha=0.5)
-
-#     # Dashed horizontal line at PAYEMS as fraction of peak equals 1
-#     fig.line(x=[-bkwd_mths_max, frwd_mths_max], y=[1.0, 1.0],
-#              color='black', line_width=2, line_dash='dashed', alpha=0.5)
-
-#     # # Create the tick marks for the x-axis and set x-axis labels
-#     # major_tick_labels = []
-#     # major_tick_list = []
-#     # for i in range(-bkwd_mths_max, frwd_mths_max + 1):
-#     #     if i % 2 == 0:  # indicates even integer
-#     #         major_tick_list.append(int(i))
-#     #         if i < 0:
-#     #             major_tick_labels.append(str(i) + 'mth')
-#     #         elif i == 0:
-#     #             major_tick_labels.append('peak')
-#     #         elif i > 0:
-#     #             major_tick_labels.append('+' + str(i) + 'mth')
-
-#     # # minor_tick_list = [item for item in range(-bkwd_mths_max,
-#     # #                                           frwd_mths_max + 1)]
-#     # major_tick_dict = dict(zip(major_tick_list, major_tick_labels))
-#     # fig.xaxis.ticker = major_tick_list
-#     # fig.xaxis.major_label_overrides = major_tick_dict
-
-#     # Add legend
-#     legend = Legend(items=[(rec_label_yrmth_lst[0], [l0]),
-#                            (rec_label_yrmth_lst[1], [l1]),
-#                            (rec_label_yrmth_lst[2], [l2]),
-#                            (rec_label_yrmth_lst[3], [l3]),
-#                            (rec_label_yrmth_lst[4], [l4]),
-#                            (rec_label_yrmth_lst[5], [l5]),
-#                            (rec_label_yrmth_lst[6], [l6]),
-#                            (rec_label_yrmth_lst[7], [l7]),
-#                            (rec_label_yrmth_lst[8], [l8]),
-#                            (rec_label_yrmth_lst[9], [l9]),
-#                            (rec_label_yrmth_lst[10], [l10]),
-#                            (rec_label_yrmth_lst[11], [l11]),
-#                            (rec_label_yrmth_lst[12], [l12]),
-#                            (rec_label_yrmth_lst[13], [l13]),
-#                            (rec_label_yrmth_lst[14], [l14])],
-#                     location='center')
-#     fig.add_layout(legend, 'right')
-
-#     # # Add label to current recession low point
-#     # fig.text(x=[12, 12, 12, 12], y=[0.63, 0.60, 0.57, 0.54],
-#     #          text=['2020-03-23', 'DJIA: 18,591.93', '63.3% of peak',
-#     #                '39 days from peak'],
-#     #          text_font_size='8pt', angle=0)
-
-#     # label_text = ('Recent low \n 2020-03-23 \n DJIA: 18,591.93 \n '
-#     #               '63\% of peak \n 39 days from peak')
-#     # fig.add_layout(Label(x=10, y=0.65, x_units='screen', text=label_text,
-#     #                      render_mode='css', border_line_color='black',
-#     #                      border_line_alpha=1.0,
-#     #                      background_fill_color='white',
-#     #                      background_fill_alpha=1.0))
-
-#     # Add title and subtitle to the plot
-#     fig_title2 = 'Progression of U.S. total nonfarm employment'
-#     fig_title3 = '(PAYEMS, seasonally adjusted) in last 15 recessions'
-#     fig.add_layout(Title(text=fig_title3, text_font_style='bold',
-#                          text_font_size='16pt', align='center'), 'above')
-#     fig.add_layout(Title(text=fig_title2, text_font_style='bold',
-#                          text_font_size='16pt', align='center'), 'above')
-
-#     # Add source text below figure
-#     updated_date_str = end_date.strftime('%B %-d, %Y')
-#     fig.add_layout(Title(text='Source: Richard W. Evans (@RickEcon), ' +
-#                               'historical PAYEMS data from FRED and BLS, ' +
-#                               'updated ' + updated_date_str + '.',
-#                          align='left',
-#                          text_font_size='3mm',
-#                          text_font_style='italic'),
-#                    'below')
-#     fig.legend.click_policy = 'mute'
-
-#     # Add the HoverTool to the figure
-#     fig.add_tools(HoverTool(tooltips=tooltips, toggleable=False,
-#                             formatters={'@Date': 'datetime'}))
-
-#     if html_show:
-#         show(fig)
-
-#     return fig, end_date_str
-
-
-# if __name__ == '__main__':
-#     # execute only if run as a script
-#     fig, end_date_str = usempl_npp()
+data_dir = os.path.join(cur_path, 'data')
+party_data_path = os.path.join(data_dir, 'deficit_party_data.csv')
+images_dir = os.path.join(cur_path, 'images')
+
+# Reading data from CVS (deficit_party_data.csv). We make six copies of the
+# DataFram and the Column Data Source to get around an error we were getting
+source=[]
+cds_list=[]
+for i in range(6):
+    df = pd.read_csv(party_data_path,
+                     dtype={'Year': np.int64,
+                            'deficit_gdp': np.float64,
+                            'receipts_gdp': np.float64,
+                            'spend_int_gdp': np.float64,
+                            'spend_nonint_gdp': np.float64,
+                            'spend_tot_gdp': np.float64,
+                            'president': 'str',
+                            'president_party': 'str',
+                            'congress_num': np.int64,
+                            'congress_sess': np.int64,
+                            'dem_whitehouse': np.int64,
+                            'dem_senateseats': np.int64,
+                            'rep_senateseats': np.int64,
+                            'oth_senateseats': np.int64,
+                            'tot_senateseats': np.int64,
+                            'dem_houseseats': np.int64,
+                            'rep_houseseats': np.int64,
+                            'oth_houseseats': np.int64,
+                            'tot_houseseats': np.int64},
+                     skiprows=3)
+    source.append(df)
+    cds_list.append(ColumnDataSource(source[i]))
+
+def output(title, file_name):
+    fig_path = os.path.join(images_dir, file_name)
+    output_file(fig_path, title=title)
+
+
+def generateFigures(title, x_label, y_label, min_seats, max_seats, min_y,
+                    max_y):
+    r'''
+    Returns a list of three empty figures for a given plot type with property
+    values to be used as the figures are filled in. This figures will become
+    the three tabs for the unified three-tab unified figure.
+
+    Args:
+        title (string): title text for figures
+        x_label (string): x-axis label for figures
+        y_label (string): y-axis label for figures
+        min_seats (scalar): minimum number of Senate or House seats in the data
+        max_seats (scalar): maximum number of Senate or House seats in the data
+        min_y (scalar): minimum y-data value
+        max_y (scalar): maximum y-data value
+
+    Returns:
+        fig_list (list): three-element list of blank Bokeh figures
+
+    '''
+    fig_list=[]
+    for i in range(3):
+        x_spread = max_seats - min_seats
+        x_buffer = 0.1
+        y_spread = max_y - min_y
+        y_buffer = 0.1
+        fig = figure(title=title,
+                     plot_height=600,
+                     plot_width=1000,
+                     x_axis_label=x_label,
+                     x_range=(min_seats - x_buffer * x_spread,
+                              max_seats + x_buffer * x_spread),
+                     y_axis_label=y_label,
+                     y_range=(min_y - y_buffer * y_spread,
+                              max_y + y_buffer * y_spread),
+                     tools=['zoom_in', 'zoom_out', 'box_zoom', 'pan', 'undo',
+                            'redo', 'reset'],
+                     toolbar_location='right')
+
+        # Set title font size and axes font sizes
+        fig.title.text_font_size = '16pt'
+        fig.xaxis.axis_label_text_font_size = '12pt'
+        fig.xaxis.major_label_text_font_size = '12pt'
+        fig.yaxis.axis_label_text_font_size = '12pt'
+        fig.yaxis.major_label_text_font_size = '12pt'
+        # Modify tick intervals for X-axis and Y-axis
+        fig.xaxis.ticker=SingleIntervalTicker(interval=10, num_minor_ticks=0)
+        fig.xgrid.ticker=SingleIntervalTicker(interval=10)
+        fig.yaxis.ticker=SingleIntervalTicker(interval=5, num_minor_ticks=0)
+        fig.ygrid.ticker=SingleIntervalTicker(interval=10)
+        # Remove Logo from toolbar
+        fig.toolbar.logo = None
+
+        fig_list.append(fig)
+
+    return fig_list
+
+
+def plotCircle(i, x_value, y_value, fig, color, src):
+    if(color == 'red'):
+        LEGEND_LABEL = 'Republican Control'
+    elif(color == 'blue'):
+        LEGEND_LABEL = 'Democrat Control'
+    else:
+        LEGEND_LABEL = 'Split Control'
+
+    fig.circle(x=source[src][x_value][i],
+               y=source[src][y_value][i],
+               size=10,
+               line_width=1,
+               line_color='black',
+               fill_color=color,
+               alpha=0.7,
+               muted_alpha=0.1,
+               legend_label = LEGEND_LABEL)
+
+
+def deficitPartyPlots(deficit_component, seat_type, src):
+    r'''
+    Generates one of six different plot types of U.S. deficit/GDP by year, by
+    Democrat held Senate seats or House seats, and by three different measures
+    of political control
+
+    Args:
+        deficit_component (string): either "deficit", "spending", or "revenues"
+        seat_type (string): either "house" or "senate"
+
+    Returns:
+        Y (array_like): aggregate output
+
+    '''
+    # Check input
+    if (not (deficit_component == 'deficit' or
+             deficit_component == 'spending' or
+             deficit_component == 'revenues')
+        or not (seat_type == 'house' or seat_type == 'senate')):
+
+        print('invalid input')
+        return
+
+    # Define variables for plot
+    data_length = len(source[src]['Year'])
+    starting_index = 0
+    if seat_type == 'house':
+        half_line = 217
+        x_label='Number of Democratic House Seats (out of 435)'
+        x_value = 'DemHouseSeats'
+        min_seats = source[src][x_value].min()
+        max_seats = source[src][x_value].max()
+        footnotes = '217 House seats'
+        if deficit_component == 'deficit':
+            fig_title = ('U.S. Federal Deficits as Percent of GDP by ' +
+                         'Democrat House Seats: 1929-2020')
+            file_name = 'deficitGDP_HouseSeats.html'
+            y_label = 'Deficit / GDP'
+            y_value = 'deficit_gdp'
+            min_y = source[src][y_value].min()
+            max_y = source[src][y_value].max()
+        elif deficit_component == 'spending':
+            fig_title = ('U.S. Federal Noninterest Spending as Percent of ' +
+                         'GDP by Democrat House Seats: 1929-2020')
+            file_name = 'spendingGDP_HouseSeats.html'
+            y_label = 'Noninterest Spending / GDP'
+            y_value = 'spend_nonint_gdp'
+            min_y = source[src][y_value].min()
+            max_y = source[src][y_value].max()
+            starting_index = 11
+        elif deficit_component == 'revenues':
+            fig_title = ('U.S. Federal Receipts as Percent of GDP by ' +
+                         'Democrat House Seats: 1929-2020')
+            file_name = 'revenuesGDP_HouseSeats.html'
+            y_label = 'Receipts / GDP'
+            y_value = 'receipts_gdp'
+            min_y = source[src][y_value].min()
+            max_y = source[src][y_value].max()
+    elif seat_type == 'senate':
+        half_line=50
+        x_label = 'Number of Democratic Senate Seats (out of 100)'
+        x_value = 'DemSenateSeats'
+        min_seats = source[src][x_value].min()
+        max_seats = source[src][x_value].max()
+        footnotes = '50 Senate seats'
+        if deficit_component == 'deficit':
+            fig_title = ('U.S. Federal Deficits as Percent of GDP by ' +
+                         'Democrat Senate Seats: 1929-2020')
+            file_name = 'deficitGDP_SenateSeats.html'
+            y_label = 'Deficit / GDP'
+            y_value = 'deficit_gdp'
+            min_y = source[src][y_value].min()
+            max_y = source[src][y_value].max()
+        elif deficit_component=='spending':
+            fig_title = ('U.S. Federal Noninterest Spending as Percent of ' +
+                         'GDP by Democrat Senate Seats: 1929-2020')
+            file_name = 'spendingGDP_SenateSeats.html'
+            y_label = 'Noninterest Spending / GDP'
+            y_value = 'spend_nonint_gdp'
+            min_y = source[src][y_value].min()
+            max_y = source[src][y_value].max()
+            starting_index = 11
+        elif deficit_component=='revenues':
+            fig_title = ('U.S. Federal Receipts as Percent of GDP by ' +
+                         'Democrat Senate Seats: 1929-2020')
+            file_name = 'revenuesGDP_SenateSeats.html'
+            y_label = 'Receipts / GDP'
+            y_value = 'receipts_gdp'
+            min_y = source[src][y_value].min()
+            max_y = source[src][y_value].max()
+
+    # Console start notification
+    print('Generating ' + file_name)
+
+    # Generate Figures
+    fig_list = []
+    fig_list = generateFigures(fig_title, x_label, y_label, min_seats,
+                               max_seats, min_y, max_y)
+
+    # Vertical black line noting half of senate seats
+    halfLine = Span(location=half_line, dimension='height', line_color='black',
+                    line_width=2)
+    for i in range(3):
+        fig_list[i].add_layout(halfLine)
+
+    # Plot data points
+    for n in range(3):
+        for i in range(starting_index, data_length):
+            if n == 0:
+                if(source[src]["DemHouseSeats"][i] < 217 and source[src]["DemSenateSeats"][i] < 50 and source[src]["DemWhitehouse"][i] == 0):
+                    plotCircle(i,x_value, y_value, fig_list[n], 'red',src)
+                elif (source[src]["DemHouseSeats"][i] > 217 and source[src]["DemSenateSeats"][i] > 50 and source[src]["DemWhitehouse"][i] == 1):
+                    plotCircle(i,x_value, y_value, fig_list[n], 'blue',src)
+                else:
+                    plotCircle(i,x_value, y_value, fig_list[n], 'green',src)
+            elif(n==1):
+                if(source[src]["DemSenateSeats"][i] < 50 and source[src]["DemWhitehouse"][i] == 0):
+                    plotCircle(i,x_value, y_value, fig_list[n], 'red',src)
+                elif (source[src]["DemSenateSeats"][i] > 50 and source[src]["DemWhitehouse"][i] == 1):
+                    plotCircle(i,x_value, y_value, fig_list[n], 'blue',src)
+                else:
+                    plotCircle(i,x_value, y_value, fig_list[n], 'green',src)
+            elif(n==2):
+                if(source[src]["DemHouseSeats"][i] < 217 and source[src]["DemWhitehouse"][i] == 0):
+                    plotCircle(i,x_value, y_value, fig_list[n], 'red',src)
+                elif (source[src]["DemHouseSeats"][i] > 217 and source[src]["DemWhitehouse"][i] == 1):
+                    plotCircle(i,x_value, y_value, fig_list[n], 'blue',src)
+                else:
+                    plotCircle(i,x_value, y_value, fig_list[n], 'green',src)
+
+    # Set up hover tool for each figure
+    TOOLTIPS = [('Year', '@Year'),
+                ('Deficit over GDP', '@deficit_gdp{0.0}'+'%'),
+                ('President','@President'),
+                ('White House', '@PresidentParty'),
+                ('Rep. House Seats', '@RepHouseSeats'),
+                ('Dem. House Seats', '@DemHouseSeats'),
+                ('Rep. Senate Seats', '@RepSenateSeats'),
+                ('Dem. Senate Seats', '@DemSenateSeats')]
+    for i in range(3):
+        fig_list[i].scatter(x=x_value, y=y_value, source=cds_list[src],
+                size=20, alpha=0, name='hover_trigger')
+        fig_list[i].add_tools(HoverTool(tooltips=TOOLTIPS, names=['hover_trigger']))
+
+   # Set up legend for each figure
+    for i in range(3):
+        fig_list[i].legend.location = 'bottom_right'
+        fig_list[i].legend.border_line_width = 2
+        fig_list[i].legend.border_line_color = 'black'
+        fig_list[i].legend.border_line_alpha = 1
+        fig_list[i].legend.label_text_font_size = '4mm'
+        # Set legend muting click policy
+        fig_list[i].legend.click_policy = 'mute'
+
+    # Add notes below plot
+    for i in range(3):
+        #Add notes below image
+        note_text_1 = ('Note: Republican control in a given year is defined as the ' +
+                    'President being Republican and Republicans holding more ' +
+                    'than '+footnotes+' for the majority of that year.')
+        caption1 = Title(text=note_text_1, align='left', text_font_size='4mm',
+                        text_font_style='italic')
+        fig_list[i].add_layout(caption1, 'below')
+        note_text_2 = ('   Democrat control is defined as the President being ' +
+                    'Democrat and Democrats holding more than '+footnotes+
+                    ' for the majority of that year. Split government is')
+        caption2 = Title(text=note_text_2, align='left', text_font_size='4mm',
+                        text_font_style='italic')
+        fig_list[i].add_layout(caption2, 'below')
+        note_text_3 = ('   defined as one party holding the White ' +
+                    'House while the other party holds a majority of House seats.')
+        caption3 = Title(text=note_text_3, align='left', text_font_size='4mm',
+                        text_font_style='italic')
+        fig_list[i].add_layout(caption3, 'below')
+        note_text_4 = ('Source: Federal Reserve Economic Data (FRED, FYFRGDA188S), ' +
+                    'United States House of Representatives History, Art, & ' +
+                    'Archives, "Party Divisions of the House of')
+        caption4 = Title(text=note_text_4, align='left', text_font_size='4mm',
+                        text_font_style='italic')
+        fig_list[i].add_layout(caption4, 'below')
+        note_text_5 = ('   Representatives, 1789 to present", '+
+                    'https://history.house.gov/Institution/Party-Divisions/Party-Divisions/, '+
+                    'Richard W. Evans (@rickecon).')
+        caption5 = Title(text=note_text_5, align='left', text_font_size='4mm',
+                        text_font_style='italic')
+        fig_list[i].add_layout(caption5, 'below')
+
+    # Output figures to an HTML file
+    output(fig_title, file_name)
+
+    panel_list=[]
+    title_str_full = 'Full control: White House + Senate + House'
+    title_str_senate = 'Senate control: White House + Senate'
+    title_str_house = 'House control: White House + House'
+    panel_list.append(Panel(child=fig_list[0], title=title_str_full))
+    panel_list.append(Panel(child=fig_list[1], title=title_str_senate))
+    panel_list.append(Panel(child=fig_list[2], title=title_str_house))
+
+    save(Tabs(tabs=panel_list))
+    # Console start notification
+    print(file_name + ' Complete')
+
+
+if __name__ == "__main__":
+    '''
+    Execute all six plots if user runs the module as a script
+    '''
+    deficitPartyPlots('deficit', 'senate', 0)
+    deficitPartyPlots('revenues', 'senate', 1)
+    deficitPartyPlots('spending', 'senate', 2)
+    deficitPartyPlots('deficit', 'house', 3)
+    deficitPartyPlots('revenues', 'house', 4)
+    deficitPartyPlots('spending', 'house', 5)
